requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
discord.py==2.2.3
exceptiongroup==1.1.1
fastf1==3.0.4
fonttools==4.39.4
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-cache==1.0.1
scipy==1.10.1
six==1.16.0
thefuzz==0.19.0
timple==0.1.6
typing_extensions==4.6.3
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
